Fix passfail crash regression
It was caused by placing a link to criteria list into engine config,
that was later populated with FailCriteria objects (that are not
serializable).

When Engine was trying to dump configuration to artifacts at the end of
prepare() stage — it couldn't serialize these objects and failed with an
pretty obscure message.