Allow a dictionary where get does not assign a value internally
Currently a value is assigned in the use of get of the dictionaries,
this causes inconveniences for those who need to use the get in a
traditional way.
The current mechanism would allow those who use it in the traditional
way to have an alternative without having to be castrating the
dictionary, and this would not allow breaking the current logic
programmed around these dictionaries.
Evaluate in the future if the anti-pattern used in the dictionary where
a value set is made internally if someone requests a get should not be
used in another method other than get.